BACKPORT: HID: nintendo: prevent needless queueing of the rumble worker
This patch adds a check for if the rumble queue ringbuffer is empty
prior to queuing the rumble workqueue. If the current rumble setting is
using a non-zero amplitude though, it will queue the worker anyway. This
is because the controller will automatically disable the rumble effect
if it isn't "refreshed".

This change improves bluetooth communication reliability with the
controller, since it reduces the amount of traffic.

Note that we still send a few periodic zero packets to avoid scenarios
where the controller fails to process the zero amplitude packet. Without
sending a few to be sure, the rumble could get stuck on until the
controller times out.

Signed-off-by: Daniel J. Ogorchock <djogorchock@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>

(cherry picked from commit dad74e18f72a852ae40ad7b4246841a1b7e196b5)

Conflicts:
   drivers/hid/hid-nintendo.c

BUG=b:236756495
TEST=Manual test with Nintendo Gamepad controller

Signed-off-by: Nergi Rahardi <nergi@google.com>
Change-Id: I262ffeafcebd4ae746e3a6d050cf0c2451fec268
Signed-off-by: fjrXTR <fajarslebew31@gmail.com>
Signed-off-by: vlxlxlv <fajarslebew31@gmail.com>