BACKPORT: HID: playstation: support DualShock4 lightbar.
Expose the lightbar LEDs in the same manner as hid-sony through
individual LEDs for backwards compatibility reasons. There is a
slight change in LED naming to use the input device name as opposed
to the MAC address like hid-sony did. This is expected to not
cause any issues and should make the naming more compliant.

In addition set a default lightbar color based on player ID.

Change-Id: I128cd2665642098f3e847471e5f31e47505609f3
Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
[cyberknight777: Backport to android12-5.10]
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: vlxlxlv <fajarslebew31@gmail.com>