BACKPORT: HID: playstation: add DualShock4 bluetooth support.
Add support for DualShock4 in Bluetooth mode. In Bluetooth, the device
is a bit strange in that after 'calibration' it switches sending all its
input data from a basic report (only containing buttons/sticks) to an
extended report, which also contains touchpad, motion sensors and other
data. The overall design of this code is similar to the DualSense code.

Change-Id: I9f03a74887010806d8ef9ba80372e4ae87735839
Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
[cyberknight777: Backport to android12-5.10]
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: vlxlxlv <fajarslebew31@gmail.com>